% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleVar_IDM.R
\name{DoubleVar_IDM}
\alias{DoubleVar_IDM}
\title{DoubleVar_IDM}
\usage{
DoubleVar_IDM(
  formFixed,
  formRandom,
  formGroup,
  formGroupVisit,
  formSurv,
  Time,
  deltas,
  timeVar,
  data.long,
  correlated_re = FALSE,
  variability_inter_visit,
  variability_intra_visit,
  sharedtype,
  hazard_baseline,
  nb.knots.splines,
  nb_pointsGK = 15,
  formSlopeFixed = NULL,
  formSlopeRandom = NULL,
  index_beta_slope = NULL,
  S1 = 500,
  S2 = 5000,
  nproc = 1,
  clustertype = "SOCK",
  maxiter = 100,
  print.info = FALSE,
  file = "",
  epsa = 1e-04,
  epsb = 1e-04,
  epsd = 1e-04,
  binit = NULL,
  partialH = NULL
)
}
\arguments{
\item{partialH}{}
}
\description{
DoubleVar_IDM
}
