% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_llh_2var_IC_rcpp.R
\name{log_llh_2var_IC_rcpp}
\alias{log_llh_2var_IC_rcpp}
\title{log_llh_2var_IC_rcpp}
\usage{
log_llh_2var_IC_rcpp(
  param,
  hazard_baseline_01,
  sharedtype_01,
  hazard_baseline_02,
  sharedtype_02,
  hazard_baseline_12,
  sharedtype_12,
  ord.splines,
  nb.beta,
  Zq,
  nb_pointsGK,
  nb.e.a,
  S,
  wk,
  rep_wk,
  sk_GK,
  nb.alpha,
  variability_inter_visit,
  variability_intra_visit,
  correlated_re,
  Case1,
  Case1bis,
  Case2,
  Case3,
  nbCase1,
  nbCase1bis,
  nbCase2,
  nbCase3,
  left_trunc,
  knots.hazard_baseline.splines_01,
  knots.hazard_baseline.splines_02,
  knots.hazard_baseline.splines_12,
  index_beta_slope,
  control
)
}
\arguments{
\item{control}{}
}
\description{
log_llh_2var_IC_rcpp
}
